+++ examples/extensive_kitchen_terraform/test/integration/extensive_suite/controls/operating_system.rb
@@ -1,15 +1,23 @@
 # frozen_string_literal: true
 
-instances_ami_operating_system_name = attribute(
-  "instances_ami_operating_system_name",
-  description: "The Terraform configuration under test must define the " \
-  "equivalently named output",
-)
+# Copyright 2016 New Context Services, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 control "operating_system" do
   desc "Verifies the name of the operating system on the targeted host"
 
   describe os.name do
-    it { should eq instances_ami_operating_system_name }
+    it { should eq attribute("instances_ami_operating_system_name") }
   end
 end