+++ examples/extensive_kitchen_terraform/test/integration/extensive_suite/controls/reachable_other_host.rb
@@ -1,15 +1,23 @@
 # frozen_string_literal: true
 
-reachable_other_host_ip_address = attribute(
-  "reachable_other_host_ip_address",
-  description: "The Terraform configuration under test must define the " \
-  "equivalently named output",
-)
+# Copyright 2016 New Context Services, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 control "reachable_other_host" do
   desc "Verifies that the other host is reachable from the current host"
 
-  describe host reachable_other_host_ip_address do
+  describe host attribute("reachable_other_host_ip_address") do
     it { should be_reachable }
   end
 end