+++ test/integration/variables/controls/variables.rb
@@ -1,15 +1,41 @@
 # frozen_string_literal: true
 
+# Copyright 2016 New Context Services, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 control "variables" do
-  describe file(::File.join(attribute("output_root_module_directory"), "string.txt")) do
+  describe "string.txt" do
+    subject do
+      file(::File.join(attribute("output_root_module_directory"), "string.txt"))
+    end
+
     its("content") { should eq "A String" }
   end
 
-  describe file(::File.join(attribute("output_root_module_directory"), "map.txt")) do
+  describe "map.txt" do
+    subject do
+      file(::File.join(attribute("output_root_module_directory"), "map.txt"))
+    end
+
     its("content") { should eq "A Value" }
   end
 
-  describe file(::File.join(attribute("output_root_module_directory"), "list.txt")) do
+  describe "list.txt" do
+    subject do
+      file(::File.join(attribute("output_root_module_directory"), "list.txt"))
+    end
+
     its("content") { should eq "Element One; Element Two" }
   end
 end